#134 - Extract remove writer
